refactor: split aria into separate files
